chore:move admin checking into controller

+++ src/permission/permission.controller.ts
@@ -1,25 +1,43 @@
 import { Controller, Get, Post, Body, Param, UseGuards } from '@nestjs/common';
 import { GetUser } from 'src/common/decorators/user.decorator';
+import { User } from '@catalog/auth/entities/user.entity';
+import { AdminGuard } from 'src/common/guards/admin.guard';
 import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
 import { PermissionService } from './services/permission.service';
 import { CreatePermissionDto } from './dto/create-permission.dto';
 import { RemovePermissionDto } from './dto/remove-permission.dto';
 import { ApiTags } from '@nestjs/swagger';
-import { User } from '@catalog/auth/entities/user.entity';
 
 @ApiTags('Permission')
 @Controller('permission')
+@UseGuards(AdminGuard)
 @UseGuards(JwtAuthGuard)
 export class PermissionController {
   constructor(private readonly permissionService: PermissionService) {}
 
-  @Post()
-  grantAccess(@Body() createPermissionDto: CreatePermissionDto) {
+  @Post('grant')
+  grantAccess(
+    @Body() createPermissionDto: CreatePermissionDto,
+    @GetUser() user: User,
+  ) {
+    if (user.isAdmin && user.id === createPermissionDto.userId) {
+      return {
+        message: `admin user already has full access to all services`,
+      };
+    }
     return this.permissionService.grantAccess(createPermissionDto);
   }
 
-  @Post()
-  removeAccess(@Body() removePermissionDto: RemovePermissionDto) {
+  @Post('remove')
+  removeAccess(
+    @Body() removePermissionDto: RemovePermissionDto,
+    @GetUser() user: User,
+  ) {
+    if (user.isAdmin && user.id === removePermissionDto.userId) {
+      return {
+        message: `cannot remove admin user from accessing a service`,
+      };
+    }
     return this.permissionService.removeAccess(removePermissionDto);
   }
   @Get()
